Handle more cases where ftdetect hangs
The first is on a Windows system, where you eventually reach the root of
a drive, and then `fnamemodify("C:", ":h:r")` returns `.`, getting stuck
in the same loop as before.

The second one is paths that don't correspond to the filesystem, such as
a `fugitive:///...` path. This eventually turns into `fugutive:/`, and
then again `.`.

This change makes it so reaching `.` is also counted as reaching root,
which will fix all these cases (as well as any others where
`fnamemodify` becomes `.`, which it appears to do if what you're doing
no longer makes sense`).

This does not supersede the previous change, as starting with the
current directory if no path is provided is still a sensible fallback.